added second seed file for sub-objects
