+++ scripts/seedDB.js
@@ -8,39 +8,7 @@ mongoose.connect(
   "mongodb://localhost/tryathlete"
 );
 
-let seedAct = {
-  sport: "Running",
-  distanceMeasurement: "50",
-  distanceUnit: "km",
-  time: 50
-}
-
-let seedBadge = {
-  isEarned: false,
-  description: "Long Run Fastly",
-  name: "Long Runner Badge",
-  sport: "Running"
-}
-
-let seedGoal = {
-  description: "Run 50 km in a week",
-  name: "First goal",
-  sport: "running",
-  isAchieved: false,
-  goalType: "distance",
-  goalQuant: 50
-}
-
-let seedChallenge = {
-  name: "First goal",
-  sport: "running",
-  isAchieved: false,
-  goalType: "distance",
-  goalQuant: 50,
-  challenger: "Jim",
-  hasAccepted: false
-}
-
+//user seed
 const userSeed = [
   {
     fname: "stephen",
@@ -76,50 +44,3 @@ db.User
     process.exit(1);
   });
 
-// seed activity test
-db.Activity.create(seedAct)
-  .then(function (dbActivity) {
-    return db.User.findOneAndUpdate({ "fname": "stephen" }, { $push: { activities: dbActivity._id } }, { new: true });
-  })
-  .then(function (dbUser) {
-    console.log(dbUser);
-  })
-  .catch(function (err) {
-    console.log(err);
-  });
-
-//seed badge object
-db.Badge.create(seedBadge)
-  .then(function (dbBadge) {
-    return db.User.findOneAndUpdate({ "fname": "stephen" }, { $push: { badges: dbBadge._id } }, { new: true });
-  })
-  .then(function (dbUser) {
-    console.log(dbUser);
-  })
-  .catch(function (err) {
-    console.log(err);
-  });
-
-//seed challenge object
-db.Challenge.create(seedChallenge)
-  .then(function (dbChallenge) {
-    return db.User.findOneAndUpdate({ "fname": "stephen" }, { $push: { challenges: dbChallenge._id } }, { new: true });
-  })
-  .then(function (dbUser) {
-    console.log(dbUser);
-  })
-  .catch(function (err) {
-    console.log(err);
-  });
-
-//seed goal object
-db.Goal.create(seedGoal)
-  .then(function (dbGoal) {
-    return db.User.findOneAndUpdate({ "fname": "stephen" }, { $push: { goals: dbGoal._id } }, { new: true });
-  })
-  .then(function (dbUser) {
-    console.log(dbUser);
-  })
-  .catch(function (err) {
-    console.log(err);
-  });
